requirements.txt
sqlalchemy>=2.0.0
pymysql>=1.1.0
pyodbc>=4.0.39
psycopg2-binary>=2.9.9
cryptography>=41.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
loguru>=0.7.0
asyncio>=3.4.3 